move files related to encoding to encode directory
