+++ packages/lib/src/offchainResolver/types.ts
@@ -1,11 +1,3 @@
-import { UserProfile } from './../account';
-//TODO add proper description and response
-export interface CiipResponse {
-    userProfile: UserProfile;
-    validUntil: number;
-    sig: string;
-}
-
 /**
  * @param name the ENS name after after decoding
  * @param record the name of the record that should be queried
